Initial commit.
Finished most of the frontend. Currently the only uncompleted feature is the ability to pass in student videos and have them be categorized in such a way that the right videos show for each question. However, the site does currently contain functionality to embed Youtube playlists if that's how we choose to implement the student videos.